requirements.txt
groq>=0.4.0
streamlit>=1.31.0
chromadb>=0.4.22
sentence-transformers>=2.3.0
tavily-python>=0.3.0
PyPDF2>=3.0.0
python-docx>=1.1.0
easyocr>=1.7.0
Pillow>=10.0.0
python-dotenv>=1.0.0
